使用例をpreview.vueに移動
