+++ src/pages/preview.vue
@@ -134,6 +134,39 @@
         ></CardCourse>
       </div>
     </section>
+    <div class="preview__content">
+      <Popup>
+        <PopupContent
+          v-for="data in popupData"
+          :key="data.value"
+          @click="data.onClick"
+          :link="data.link"
+          :value="data.value"
+          :color="data.color"
+        ></PopupContent>
+      </Popup>
+    </div>
+    <div class="preview__content">
+      <Popup>
+        <PopupContent
+          v-for="data in popupModuleData"
+          :key="data.value"
+          @click="data.onClick"
+          :value="data.value"
+        >
+        </PopupContent>
+      </Popup>
+    </div>
+    <section class="preview__content">
+      <Button @click="openModal" size="small">Open Modal</Button>
+    </section>
+    <section class="preview__content">
+      <Modal v-if="modal" @click="closeModal" size="large">
+        <div class="modal__title">title</div>
+        <div class="modal__contents">contents</div>
+        <Button @click="closeModal" layout="fill" size="medium">OK</Button>
+      </Modal>
+    </section>
   </article>
 </template>
 
@@ -153,6 +186,9 @@ import SidebarContent from "../components/SidebarContent.vue";
 import DecoratedIcon from "../components/DecoratedIcon.vue";
 import CardAdd from "../components/CardAdd.vue";
 import CardCourse, { Course } from "../components/CardCourse.vue";
+import Popup from "../components/Popup.vue";
+import PopupContent from "../components/PopupContent.vue";
+import Modal from "../components/Modal.vue";
 
 export default defineComponent({
   name: "Preview",
@@ -167,6 +203,9 @@ export default defineComponent({
     DecoratedIcon,
     CardAdd,
     CardCourse,
+    Popup,
+    PopupContent,
+    Modal,
   },
   setup: () => {
     const { ready, state } = useUsecase(authCheck, true);
@@ -196,6 +235,64 @@ export default defineComponent({
       welcomeModal.value = false;
     };
 
+    // popup
+    const popupClickHandler = (value: string) => {
+      console.log(value);
+    };
+
+    const popupData = [
+      {
+        value: "編集する",
+        link: false,
+        onClick: popupClickHandler,
+        color: "normal",
+      },
+      {
+        value: "シラバス",
+        link: true,
+        onClick: popupClickHandler,
+        color: "normal",
+      },
+      {
+        value: "出席(respon)",
+        link: true,
+        onClick: popupClickHandler,
+        color: "normal",
+      },
+      {
+        value: "地図",
+        link: true,
+        onClick: popupClickHandler,
+        color: "normal",
+      },
+      {
+        value: "削除する",
+        link: false,
+        onClick: popupClickHandler,
+        color: "danger",
+      },
+    ];
+
+    const popupModuleData = [
+      { value: "春A", onClick: popupClickHandler },
+      { value: "春B", onClick: popupClickHandler },
+      { value: "春C", onClick: popupClickHandler },
+      { value: "秋A", onClick: popupClickHandler },
+      { value: "秋B", onClick: popupClickHandler },
+      { value: "秋C", onClick: popupClickHandler },
+    ];
+
+    // modal
+    const modal = ref(false);
+
+    const openModal = () => {
+      modal.value = true;
+    };
+
+    const closeModal = () => {
+      modal.value = false;
+    };
+
     return {
       displayLog,
       isBtnActive,
@@ -207,12 +304,19 @@ export default defineComponent({
       welcomeModal,
       openWelcomeModal,
       closeWelcomeModal,
+      popupData,
+      popupModuleData,
+      modal,
+      openModal,
+      closeModal,
     };
   },
 });
 </script>
 
 <style lang="scss" scoped>
+@import "../scss/main.scss";
+
 .course-grid {
   display: grid;
   grid-template-rows: repeat(2, 1fr);
@@ -234,5 +338,43 @@ export default defineComponent({
 
 .card-course-wrapper {
   width: 40rem;
+}
+
+// modal
+.modal {
+  &__title {
+    font-size: 2rem;
+    font-weight: 500;
+    line-height: $multi-line;
+    color: $text-main;
+    text-align: left;
+  }
+  &__contents {
+    width: 100%;
+    margin: 4rem 0 2.8rem 0;
+    font-size: 1.6rem;
+    font-weight: normal;
+    line-height: $multi-line;
+    color: $text-main;
+    text-align: left;
+    overflow: scroll;
+  }
+  // コンテンツの高さ
+  &--large {
+    .modal__contents {
+      height: calc(56.7rem - 20.8rem);
+      @include large-screen {
+        height: calc(51.58334rem - 20.8rem);
+      }
+    }
+  }
+  &--small {
+    .modal__contents {
+      height: calc(39.5rem - 20.8rem);
+      @include large-screen {
+        height: calc(30.33334rem - 20.8rem);
+      }
+    }
+  }
 }
 </style>
